Fix version checks for 4.7.x kernels
Checking for "4.6.0 or less" when trying to determine when to use 4.7.x
series features does the wrong thing when running a 4.6 kernel when it
isn't the first release; a version of 4.6.4 will cause the selection of
4.7.x features, which will proceed to fail to build because the kernel
is a 4.6.x series. Fix the condition so that the right thing happens
with later 4.6.x series kernels.