Make painter_[un]lock log from where it was called
EVDI_CHECKPT logs current function name and line in file with __func__
and __LINE__. This makes it log always the same function name
(painter_lock/unlock) and the same line in file. With this we only know
that mutex is locked or unlocked, but we don't know which funcion called
it.

Replacing function with macro inlines code into the caller function so
both __func__ and __LINE__ will be pointer to 'where' it was called.