Don't dereference pointer state if it is null
An earlier check on the pointer 'state' implies it could be null
and the calls drm_atomic_state_put/drm_atomic_state_free can
dereference it causing a null pointer dereference if it is null.
Add a null check to state so that we avoid the null dereferences.

Issue found by static analysis checking with Coverity.

Addreses-Coverity: ("Dereference after null check")
Signed-off-by: Colin Ian King <colin.king@canonical.com>