Do not drop reference to drm_device in evdi gem release function
This put_device was introduced in commit:
https://github.com/torvalds/linux/commit/347cf10 in times of v3.14 kernels.
Most likely this call was unnecessary in that line. This was mainly done for udl driver for usb unplug.
This is not the case for evdi driver.

What is more removing reference to drm_device in drm_gem_object free function is wrong.