Clarify what needs admin rights
This change moves the warning about admin rights to the calls that
require it.

My understanding is writing to the sysfs requires admin rights, but the
ioctls don't. Given this, it is possible to ensure an app doesn't need
admin rights by either configuring the kernel module or having a
separate program that adds devices.